Use local modules instead of environment variables
