--- app/trustpilotApi.js
@@ -1,67 +1,0 @@
-'use strict';
-
-// Allow dependency injection of an extended client module
-const Trustpilot = process.env.TRUSTPILOT_CLIENT_MODULE ?
-  require(process.env.TRUSTPILOT_CLIENT_MODULE) :
-  require('trustpilot');
-
-module.exports = function (config) {
-
-  const client = new Trustpilot({
-    apiKey: config.API_KEY,
-    secret: config.API_SECRET,
-    username: config.BUSINESS_USER_NAME,
-    password: config.BUSINESS_USER_PASS,
-    baseUrl: config.API_HOST,
-  });
-  const BUSINESS_UNIT_ID = config.BUSINESS_UNIT_ID;
-
-  const ApiBridge = (() => {
-
-    async function privateRequest(options) {
-      const requestWrapper = await client.authenticate();
-      return requestWrapper(options);
-    }
-
-    return {
-      getLastUnansweredReview: async function ({
-        stars,
-        businessUnitId,
-      }) {
-        const params = {
-          orderBy: 'createdat.desc',
-          responded: false,
-        };
-        if (stars) {
-          params.stars = stars;
-        }
-        businessUnitId = businessUnitId || BUSINESS_UNIT_ID;
-        const data = await privateRequest({
-          method: 'GET',
-          uri: `/v1/private/business-units/${businessUnitId}/reviews`,
-          qs: params,
-        });
-        if (data.reviews.length > 0) {
-          return data.reviews[0];
-        }
-
-        return null;
-      },
-
-      replyToReview: function ({
-        reviewId,
-        message,
-      }) {
-        return privateRequest({
-          method: 'POST',
-          uri: `/v1/private/reviews/${reviewId}/reply`,
-          form: {
-            message: message,
-          },
-        });
-      },
-    };
-  })();
-
-  return ApiBridge;
-};
